organization_service/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
kafka-python==2.1.0